% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia.R
\name{cia}
\alias{cia}
\title{Insert animated image. This is still in production.}
\usage{
cia(src, alt = "", height = "100\%", width = "100\%", size = 1,
  position = c(0, 0, 0, 0), grow = 0, animate = "flash",
  speed = "normal", anitype = "repeat", rotate = 0, flip = "none",
  border = FALSE, bgcolor = NULL, bgcolour = bgcolor, iother = NULL,
  sother = NULL, dother = NULL)
}
\arguments{
\item{src}{A url or path of image to animate.}

\item{alt}{Alternative text to the image.}

\item{height}{height of the image that can be specified in percentage or pixels.}

\item{width}{width of the image that can be specified in percentage or pixels.}

\item{position}{A numerical vector of length 4 specifying the value to move up, down, left and right.}

\item{grow}{Numerical value to grow (or shrink for negative values) without changing or moving the container.}

\item{animate}{'bounce', 'burst', 'falling', 'flash', 'float', 'horizontal',
'passing', 'passing-reverse', 'pulse', 'ring', 'shake', 'spin',
'tada', 'vertical', or 'wrench'}

\item{speed}{'normal', 'fast', or 'slow'}

\item{anitype}{'repeat' or 'hover'}

\item{rotate}{Numerical value for degree of rotation. This is not working.}

\item{flip}{'none', 'horizontal', 'vertical'.}

\item{border}{If TRUE, draws a border around the icon.}

\item{bgcolor, bgcolour}{Colour to be given to the background}

\item{iother}{Character vector of other parameters directly added to the icon classes, e.g. fa-pull-left, fa-pull-right.}

\item{sother}{Character vector of other parameters directly added to the style classes.}

\item{dother}{Character vector of other parameters directly added to the data-fa-transform classes, e.g. shrink-8, grow-2}

\item{align}{top, bottom, middle, left, right}
}
\description{
Insert animated image. This is still in production.
}
\references{
\href{https://l-lin.github.io/font-awesome-animation/}{Font awesome animation}
}
